# openapi/schema.yaml
openapi: 3.0.3
info:
  title: PharmaPocket API
  version: 1.0.0
  description: Contrat de l'API v1 consommée par le frontend PharmaPocket.
paths:
...
  /api/v1/content/admin/images/upload/:
    post:
      operationId: admin_image_upload
      tags:
      - content
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminImageUpload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminImageUpload'
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagePayload'
          description: ''
  /api/v1/content/admin/microarticles/search/:
    get:
      operationId: admin_microarticle_search
      parameters:
      - in: query
        name: maladies_nodes
        schema:
          type: string
      - in: query
        name: medicament_nodes
        schema:
          type: string
      - in: query
        name: pharmacologie_nodes
        schema:
          type: string
      - in: query
        name: q
        schema:
          type: string
      - in: query
        name: recent
        schema:
          type: boolean
      - in: query
        name: tags
        schema:
          type: string
      - in: query
        name: theme_nodes
        schema:
          type: string
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminMicroArticleSearchResult'
          description: ''
  /api/v1/content/admin/packs/:
    get:
      operationId: admin_pack_list
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminPackSummary'
          description: ''
    post:
      operationId: admin_pack_create
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminPackCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminPackCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackSummary'
          description: ''
  /api/v1/content/admin/packs/{pack_id}/:
    get:
      operationId: admin_pack_retrieve
      parameters:
      - in: path
        name: pack_id
        schema:
          type: integer
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackDetail'
          description: ''
    patch:
      operationId: admin_pack_update
      parameters:
      - in: path
        name: pack_id
        schema:
          type: integer
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAdminPackPatch'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAdminPackPatch'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAdminPackPatch'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackSummary'
          description: ''
    delete:
      operationId: admin_pack_delete
      parameters:
      - in: path
        name: pack_id
        schema:
          type: integer
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/content/admin/packs/{pack_id}/bulk-add/:
    post:
      operationId: admin_pack_bulk_add
      parameters:
      - in: path
        name: pack_id
        schema:
          type: integer
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackBulkAdd'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminPackBulkAdd'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminPackBulkAdd'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAddResponse'
          description: ''
  /api/v1/content/admin/packs/{pack_id}/cards/{card_id}/remove/:
    post:
      operationId: admin_pack_card_remove
      parameters:
      - in: path
        name: card_id
        schema:
          type: integer
        required: true
      - in: path
        name: pack_id
        schema:
          type: integer
        required: true
      tags:
      - content
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
          description: ''
  /api/v1/content/admin/packs/{pack_id}/cards/reorder/:
    post:
      operationId: admin_pack_card_reorder
      parameters:
      - in: path
        name: pack_id
        schema:
          type: integer
        required: true
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackReorder'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminPackReorder'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminPackReorder'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountUpdateResponse'
          description: ''
...
  /api/v1/tags/:
    get:
      operationId: tag_list
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
      - in: query
        name: q
        schema:
          type: string
      tags:
      - tags
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagPayload'
          description: ''
...
components:
  schemas:
...
    AdminImageUpload:
      type: object
      description: |-
        POST /admin/images/upload/ — le fichier arrive sous `file` ou `image`.

        `validated_data` expose `file` (l'upload validé) et `title` (celui fourni,
        sinon le nom du fichier).
      properties:
        file:
          type: string
          format: binary
        image:
          type: string
          format: binary
        title:
          type: string
          nullable: true
    AdminMicroArticleSearchResult:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        packs_count:
          type: integer
          minimum: 0
      required:
      - id
      - packs_count
      - slug
      - title
    AdminPackBulkAdd:
      type: object
      description: |-
        POST /admin/packs/<id>/bulk-add/.

        Trois formes d'entrée, dans cet ordre de priorité : `items` (texte collé),
        `microarticle_ids` puis `slugs`. Toutes se ramènent à
        `validated_data["tokens"]`, une liste de jetons id-ou-slug que la vue résout.
      properties:
        items:
          type: string
        microarticle_ids:
          type: array
          items:
            type: string
        slugs:
          type: array
          items:
            type: string
    AdminPackCreate:
      type: object
      description: POST /admin/packs/ — `sort_order` absent ou nul = à la suite des
        packs existants.
      properties:
        name:
          type: string
          maxLength: 60
        description:
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
          maxLength: 16
        estimated_minutes:
          type: integer
          maximum: 32767
          minimum: 0
          nullable: true
        status:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/StatusEnum'
          - $ref: '#/components/schemas/NullEnum'
        cover_image_id:
          type: integer
          nullable: true
        sort_order:
          type: integer
          nullable: true
      required:
      - name
    AdminPackDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
        estimated_minutes:
          type: integer
          nullable: true
        status:
          type: string
        type:
          type: string
        sort_order:
          type: integer
        cards_count:
          type: integer
          minimum: 0
        cover_image_url:
          type: string
          nullable: true
        cover_image_credit:
          type: string
          nullable: true
        cover_image:
          allOf:
          - $ref: '#/components/schemas/ImagePayload'
          nullable: true
        cards:
          type: array
          items:
            $ref: '#/components/schemas/DeckCardItem'
      required:
      - cards
      - cards_count
      - cover_image_url
      - description
      - difficulty
      - estimated_minutes
      - id
      - name
      - sort_order
      - status
      - type
    AdminPackReorder:
      type: object
      description: POST /admin/packs/<id>/cards/reorder/
      properties:
        microarticle_ids:
          type: array
          items:
            type: integer
      required:
      - microarticle_ids
    AdminPackSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
        estimated_minutes:
          type: integer
          nullable: true
        status:
          type: string
        type:
          type: string
        sort_order:
          type: integer
        cards_count:
          type: integer
          minimum: 0
        cover_image_url:
          type: string
          nullable: true
        cover_image_credit:
          type: string
          nullable: true
        cover_image:
          allOf:
          - $ref: '#/components/schemas/ImagePayload'
          nullable: true
      required:
      - cover_image_url
      - description
      - difficulty
      - estimated_minutes
      - id
      - name
      - sort_order
      - status
      - type
...
    BulkAddResponse:
      type: object
      properties:
        added:
          type: integer
          minimum: 0
        already_present:
          type: integer
          minimum: 0
        not_found:
          type: integer
          minimum: 0
      required:
      - added
      - already_present
...
    CardType:
      enum:
      - standard
      - recap
      - detail
      type: string
      description: |-
        * `standard` - standard
        * `recap` - recap
        * `detail` - detail
    CategoryPayload:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
...
    CountUpdateResponse:
      type: object
      properties:
        ok:
          type: boolean
        updated:
          type: integer
          minimum: 0
      required:
      - ok
      - updated
...
    DeckCardItem:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        answer_express:
          type: string
        takeaway:
          type: string
        key_points:
          type: array
          items:
            type: string
        cover_image_url:
          type: string
          nullable: true
        cover_image_credit:
          type: string
          nullable: true
        cover_image:
          allOf:
          - $ref: '#/components/schemas/ImagePayload'
          nullable: true
        tags:
          type: array
          items:
            type: string
        tags_payload:
          type: array
          items:
            $ref: '#/components/schemas/TagPayload'
        categories_theme_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        categories_maladies_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        categories_medicament_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        categories_pharmacologie_payload:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPayload'
        published_at:
          type: string
          format: date-time
          nullable: true
        decks_count:
          type: integer
          minimum: 0
        card_type:
          $ref: '#/components/schemas/CardType'
        subject:
          allOf:
          - $ref: '#/components/schemas/SubjectSummary'
          nullable: true
        deck_card_id:
          type: integer
        position:
          type: integer
        sort_order:
          type: integer
        is_optional:
          type: boolean
        notes:
          type: string
      required:
      - answer_express
      - cover_image_url
      - id
      - key_points
      - slug
      - tags
      - takeaway
      - title
...
    ImagePayload:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
          nullable: true
        credit_text:
          type: string
          nullable: true
        credit_author:
          type: string
        credit_source:
          type: string
        credit_source_url:
          type: string
        credit_license:
          type: string
        credit_license_url:
          type: string
      required:
      - id
      - title
      - url
...
    NullEnum:
      enum:
      - null
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
      - ok
...
    PatchedAdminPackPatch:
      type: object
      description: PATCH /admin/packs/<id>/
      properties:
        name:
          type: string
          maxLength: 60
        description:
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
          maxLength: 16
        estimated_minutes:
          type: integer
          maximum: 32767
          minimum: 0
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        cover_image_id:
          type: integer
          nullable: true
        sort_order:
          type: integer
...
    StatusEnum:
      enum:
      - draft
      - published
      - archived
      type: string
      description: |-
        * `draft` - Draft
        * `published` - Published
        * `archived` - Archived
...
    SubjectSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
      required:
      - description
      - id
      - name
      - slug
    TagPayload:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid